[refactor] devide presentation-container by using emits event handling
